Manager's Interface Refinements

+++ src/test/java/ManagerController.java
@@ -116,7 +116,7 @@ class TestManagerController {
 
     @Test
     void addDishShouldFailWhenPriceLessThanZero() {
-        testDish.setPrice(-1);
+        testDish.setPrice((float) -1);
         EmptyResponse response = controller.addDish(testDish);
 
         assertFalse(response.isSuccess());
